feat(rate-limiting-advanced): add `config.disable_penalty` to not counting the rejected 429 requests (#3762)
The current implementation always increments the counter whether the request is rejected or not. So if the request rate exceeds the configured quota, the requests will always be rejected with 429.

The `config disable_penalty` is added to control this behavior. When it's enabled, the rejected 429 requests won't be counted. And the default value of disable_penalty is false, which keeps the default behavior unchanged.

FTI-3281