tests(helpers): use io.open to truncate file
The old implementation of clean_logfile() exec()-ed out to /bin/sh with
redirection to truncate the file. This yields an annoying error message
if the file does not exist and cannot be created because its parent
directory also doesn't exist:

> sh: 1: cannot create /home/runner/work/kong-ee/kong-ee/servroot/logs/error.log: Directory nonexistent

This implementation tries to be a little smarter and also avoids exec().