fix(dbless) always applying license before entity schema verification
Applying a configuration to Kong Gateway in hybrid mode can cause issues
for newly connected data planes where the payload being processed has
validation requiring a license applied. This was issue introduced with
[kong-ee:f8f39c3][1] where enterprise only vaults are being restricted
to licensed customers. Because the license is not applied before the
rest of the payload is processed/validated, thus a “bad config” is
possible for these initial payloads sent to the Kong Gateway data plane
when a license is not set via env/file during startup.

Actually, license takes effect is done asynchronously via
[event_handler][2]. The fix here is applying the license before the
schema verification phase when using declarative configuration.

[1]:
https://github.com/Kong/kong-ee/commit/f8f39c304774783db6289cf2e678fe79c27089da
[2]:
https://github.com/Kong/kong-ee/blob/b7627d7473fa21257f5b999348569623f94a714a/kong/enterprise_edition/licensing.lua#L143

Fix FTI-4502