fix(rate-limiting) bypass strategy init if strategy_type is off or sync_rate is -1 (#3822)
If `strategy_type` is `off`, we should not load it, otherwise error
would be reported as:

```
2022/08/24 14:13:40 [error] 29717#0: *1 [kong] handler.lua:119 [graphql-rate-limiting-advanced] err in creating new ratelimit namespace:
./kong/tools/public/rate-limiting/init.lua:504: module 'kong.tools.public.rate-limiting.strategies.off' not found:
No LuaRocks module found for kong.tools.public.rate-limiting.strategies.off
```

If `sync_rate` is set to `-1`, we will not load it as well, for perf
improvement.

We return earlier in functions like `sync`, `fetch`, etc. to improve
perf.