fix(*): remove /tmp/build prefix from paths returned by debug.getinfo (#4143)
Due to the way how we precompile EE plugins for shipping, their
debuginfo contains /tmp/build as path prefix.  This change corrects
these paths so that they point to the files that are actually
installed.  This benefits the datafile plugin, which uses debuginfo
to locate auxilary files installed by luarocks.