feat(aws-lambda): Allow IMDSv2 as protocol to retrieve EC2 instance role (#9962)
This adds optional support for the IMDSv2 protocol.  By default, the
aws-lambda plugin will use the IMDSv1 protocol to retrieve the
instance role.  Users can configure it to use IMDSv2 instead by
setting the configuration parameter `aws_imds_protocol_version` to
`v2`.  The default for the parameter is `v1` to ensure that
installations that run Kong inside of Docker on EC2 will continue to
work unchanged.

Signed-off-by: Ivan Savcic <isavcic@gmail.com>
Co-authored-by: Ivan Savcic <ivans@vast.com>