fix(plugins/datadog):  add return value for log function in batch queue (#10044)
The log function in Datadog is called by the batch
queue when a batch is processed, and batch queue relys
on the return value of the callback. The return value of log
function in Datadog always return `nil`, this makes batch queue consider
the result of processing as failed. In this commit, the correct return value
indicating the success in processing to fix this bug.