tests(websockets): increase timeout in test case (#4301)
This test has become somewhat flaky over time, which is concerning for
its own reasons, but it should probably be addressed by raising the
timeout first.

This raises the connect timeout in an upstream/service timeout test case
from 100ms to 1000ms. It's a heavily parameterized test suite, so the
effect is somewhat dramatic: the entire test suite takes 60s-70s for me
locally, while it used to be more like 10s-15s. It's unfortunate, but
the fact remains that 100ms is pretty low for a real-world timeout. The
low value was initially chosen in order to make the tests fast. The
downside is that such a low value leaves little room for tolerating any
varience without making the test case meaningless.

I added the #slow tag to the test case, in case we want a way to isolate
this kind of test in CI or local development.

I also performed a minor code change to increase readability.