fix(dbless): trigger event to allow Vitals/events_hooks subsystem to be reinitialized during license preload (#4232)
This fixes the runtime initialization of the Vitals reports during
license preload. A `kong:configuration change` event is triggered to
start/stop the vitals reports on the dataplanes when the license is
added and the featureset reverts to full. Hence the DP is able to submit
the reports properly.

FTI-4621