feat(core): adds front requeuing
This commit proposes the optional parameter 'process_requeue_front' which adds requeued entries back to the front of the queue.
Current default behavior are that entries get requeued to the back and isn't changed when this parameter is unset or set to `false`.

Motivation:
With the default behavior e.g. used in the http-log plugin there can be situations where the queue piles up a lot of data in case entries cannot be processed since the queue has a LIFO behavior in case of errors.
In case the process error persists over several minutes the retry delay of the queue is increased to 60s. When new entries are stored in front of the queue and processed first they will also fail. This means older batches are not processed, the retry counter for these old batches won't be increased and they won't expire causing high-memory usage in the end.

With this commit a FIFO behavior is introduced ensuring, that the the retry for the oldest batch is increased first and the batch is actually dropped when max_retries is reached before processing newer records.