fix(template) keep UTF-8 as default for charset directive but make it configurable (#10111)
### Summary

`charset` directive that was previously hard coded in template with a value of `UTF-8`
was removed completely with this PR:
https://github.com/Kong/kong/pull/9905

Previous discussion from 3 years ago pointed that this is breaking change (not a fix):
https://github.com/Kong/kong/pull/5045

So here is another fix to make the directive still default to `UTF-8`. which is a quite
fine default, but allow users to configure it. E.g.:

```
KONG_NGINX_HTTP_CHARSET=off kong start
```