chore(plugins): cherry pick batch queue return value fix to 3.1.x (#10126)
* fix(plugins/datadog):  add return value for log function in batch queue (#10044)

The log function in Datadog is called by the batch
queue when a batch is processed, and batch queue relys
on the return value of the callback. The return value of log
function in Datadog always return `nil`, this makes batch queue consider
the result of processing as failed. In this commit, the correct return value
indicating the success in processing to fix this bug.

* fix(plugins/statsd & opentelemetry): add return values for batch queue callback functions (#10052)

Add return values for functions that will be processed in batch queue for these two plugins.
Batch queue considers the result of processing unsuccessful without a return
value, resulting the batch being reprocessed incorrectly.

FTI-4645

Co-authored-by: Robin Xiang <liverpool8056@163.com>