fix(ci): sanitize config comments processing and hashmark quoting
Up until now, values read from kong's configuration file could include
comments introduced by a hashmark ('octothorpe') and quoted hashmarks.
Unquoting and stripping was later done when the value was parsed
('inferred').  This caused issues when values were not read from the
configuration file, but rather from the environment or from a vault.
Values got the same stripping/unquoting treatment and to prevent that
from happening, code was added to specially process values that did
not come from the configuration file.

With this change, removal of comments and unquoting of quoted
hashmarks is moved from the value parsing code to the code that reads
the configuration file.  This removes the need for special treatment
of environment (or vault) values.

This change also fixes a bug in the previous comment stripping code
that required hashmarks introducing comments to always be preceded
with a space character that would be stripped, too.

Fixes #10098, KAG-456