fix(pdk): request.get_raw_query
Avoid accessing ngx.var.args and fetch the raw query string from the
request_uri.

Accessing ngx.var.args was causing issue #10080. We might want to
restrict access to happen solely via ngx.var.* in the future and/or to
implement some index refreshing logic in lua-kong-nginx-module but right
now this is the easiest way to avoid the problem in the pdk.