style(bug_report): Allow reproduce formatting (#10168)
Allow the "steps to reproduce" section to be formatted and previewed as
Markdown like the other sections. It's currently not clear how it will
be formatted when you're writing an issue. Everything gets put inside a
single code fence and I frequently find that I need to reformat it after
first submitting an issue.