refactor(http-log) avoid creating big tables when queuing
Without this change, `json_array_concat` creates a lot of temporary strings that will increase memory churn unnecessarily. The temp string will be bigger as the number of queued items grows.

Here we use an iterator function. This function is fixed-size and will not grow as the batch size grows.