feat(*): add datadog-tracing plugin (#4364)
Add Datadog tracing plugin. The exporter in the log phase encodes traces in the msgpack format and sends them to the Datadog Agent through HTTP. This plugin uses the tracing PDK and collects build-in tracing instrumentation like the OpenTelemetry plugin. Additional tags are added to identify Kong services and routes.

Currently, this plugin relies on two options: opentelemetry_tracing and opentelemetry_tracing_sampling_rate; these config names are considered to be changed to general ones for all tracing plugins (otel and dd). But it should be addressed on the CE side first.

KAG-241