fix(jsonschema): a 'type' field can be burried under a reference (#4417)
the validation for a parameter checks if the schema has a `type` field
at the top level. This is correct, but should take into account that
the top-level actually is a reference. In that case it needs to follow
the reference and look for the `type` field at the reference target.