feat(canary): add default value for canary start time field (#3715)
This PR adds a default value for the start time field in the canary plugin.

Currently when the hash field is setting as one_of = { "consumer", "ip", "none" }, we'll require a mandatory value for start or percentage.

This PR modifies the behaviour, after the change, if hash is not set to allow or deny(which indicates a manual control flow of the canary release), we'll give the start field a default value, which is ngx.time. It means that the canary flow will start at the next second by default. Of course, this will not influence the percentage fix step behaviour, as the percentage always takes precedence.

FTI-4281