fix(docker): add -n option to avoid ln failing to overwrite the existing symbolic link file
On some older kernels (below 4.2), executing `ln -sf /dev/stdout
/tmp/access.log` multiple times will only work the first time. This is because
those kernels have a bug that misleads `openat()` to recognize `/dev/stdout` as
a directory, leading to `ln` does not able to overwrite the existing files.

This can lead to that once the container is started, it will continue to fail
when restarting afterward due to the docker entrypoint does not execute
properly.

Ref:
[1]:
https://stackoverflow.com/questions/75305383/openat-recognized-dev-stdout-as-a-d
irectory