fix(dbless): drain error table when flattening errors
This only affect `POST /config?flatten_errors=1`.

When nested errors are flattened for the response body, they should
be removed from the original error table so that they are not
included in the final `fields` error object. Errors that are not
flattened should remain.

It could be argued that this is a backwards-incompatible change, but:

1. This is opt-in only via the `flatten_errors` query param.
2. It's arguably a behavioral fix since we are effectively
   de-duplicating the errors in the response.
3. This makes the response object much, much cleaner.