fix(*): prevent queues from growing without bounds (#10046) (#10253)
This commit implements an upper limit on the number of batches that
may be waiting on a queue for processing.  Once the limit has been
reached, the oldest batch is dropped from the queue and an error
message is logged.  The maximum number of batches that can be waiting
on a queue is configured through the max_queued_batches parameter of
the queue, which defaults to 100 and can be globally overriden with
the max_queued_batches parameter in kong.conf

A more complete rewrite of the queuing system is in the works but will
not make it into 3.2.

KAG-303

This reverts commit 218cc0a81500986b0b94e6b4e350302326097018.