tests(websocket): tag one test case as #flaky (#4529)
This one has been failing a lot lately in CI, and locally for some
users. I cannot reproduce the failure locally, so it's a royal pain
to debug. There have been no changes to this area of the code in a long
while, so I am confident that:

A) it's really just a flaky test that needs to be fixed
B) something else is amiss, but has been amiss for a long while

In any case, the the scenario that is being tested is not super duper
critical functionality, so finding+fixing the root cause can be
considered low priority.