feat(batchqueue): increase default maximum queue size and remove global configuration parameter (#10271)
We don't want to support this parameter going forward as we're going
to re-implement the queue system.  This change hard-codes the queue limit
to a value that would be considered "large enough".

We added an undocumented environment variable serves to help users who require
a different limit on the maximum size of queues.  A proper solution is
under development and this undocumented environment variable will only
exist until the reimplementation has been released

Co-authored-by: Datong Sun <datong.sun@konghq.com>