fix(api): /userinfo to return better value for session renew (#4596)
* fix(api): /userinfo to return better value for session renew

### Summary

When we bumped session library, there was new working parameter called `idling_timeout`
which now defaults to `900` (previously it defaulted to `0`). The manager and admin
also uses so called `renew` parameter which we removed from the library. The renewing
and calculating when session will expires is thus changed.

Kong Manager tests showed that manager started now popup "session expires in 15 minutes"
warnings. Previously it used `renew` value, which by default was `600` and calculated
10% from it, leading to fact that warning would popup when there were 60 secs from session
expiry.

We did not have time to workaround this on Manager side, so we workaround it on Kong side
in this commit. This commit will take minumum value of `absolute_timeout`, `rolling_timeout`
and `idling_timeout` or constant `600`. If all config settings are set to `0` (effectively
making session never expire until browser is closed) it will return renew of `0`.

So by default it will work like this:

1. user logs in manager (absolute_timeout = 1 day, rolling_timeout = 1 hour, idling_timeout = 15 mins)
2. if user does nothing the for 14 minutes, the popup will show up

If idling_timeout is configured to 300 it will use that instead of default `600`, which means that popup will be
shown 30 seconds before idling period ends. This 10% logic is inside manager, so we cannot affect that.

---------

Co-authored-by: bungle <bungle@konghq.com>
Co-authored-by: samugi <samuele@konghq.com>
(cherry picked from commit 57df0ebb275892e9bdd56d8bc239e7e75ea90b78)