tests(helpers) avoid spurious error message (#10368)
If the inner command substitution failed with a "file not found", the
error would be reported, even though the outer command has a
redirection. This happens because the substitution is evaluated first,
and the outer redirection does not apply to it. This fixes a flakiness
in the hybrid mode tests for `kong quit`. This commit also does some
minor cleanups in said tests.