perf(rbac): fix a possible slow query in rbac_role_entities deletion (#4216)
When enforce_rbac is set to entity, when we delete an entity, for
example, a route is deleted, we should ensure the related entries in
table rbac_role_entities are deleted at the same time. In original
logic, we use each() to retrieve all the entries, when the table is too
large, it takes a long time. In this PR, a custom function is made to do
this dedicated operation in a more efficient way.

FTI-4532