fix(ci): pongo to use docker image for current commit (#4427)
This commit builds a docker image using the artifact from build step, export to artifacts then download back
in pongo tests.

This commit also uses the `master` branch of pongo. When this is stablized and pongo released a new version,
the `ref` of kong-pongo should be then changed back to a tag.