fix(forward-proxy) latencies.proxy is not right when forward-proxy is enabled (#4538)
is enabled, latencies.proxy is changed to be calculated in forward-proxy
plugins followed by the httpc:connect() called where the requests
for upstreams being sent, instead of being calculated in header_filter.

Based on the fact that the `latencies.proxy` is designed to indicated
the waiting time for the first byte of response from upstream after
the request being forwarded to upstream. In the case that forward-proxy
is enabled, if following the original logic calculated in header_filter,
it is evaluated from `ctx.KONG_HEADER_FILTER_START - ctx.KONG_ACCESS_ENDED_AT`.
`ctx.KONG_ACCESS_ENDED_AT` is calculated closer to the end of access
phase, followed by `ctx.KONG_HEADER_FILTER_START` being calculated in
header_filter phase, which is caused by a callback function generated
in the handler of the plugin being called for sending response to
client at the end of access phase. It means it is the time for callback
to read the first chunk of the response.

FTI-1904