feat(opentelemetry): Allow header_type to be set (#10620)
* feat(opentelemetry): Allow header_type to be set

At the moment the opentelemtry plugin sets the `default_header_type` for propigations:

https://github.com/Kong/kong/blob/master/kong/plugins/opentelemetry/handler.lua#L150

However, doesn't allow the modification of `header_type` in the `schema`, so it always defaults to `preserve`. However, based on the function it's calling:

https://github.com/Kong/kong/blob/master/kong/tracing/propagation.lua#L432

> -- If conf_header_type is set to `preserve`, found_header_type is used over default_header_type;

The above means the `default_header_type` is never used. We want the `default_header_type` to be used, and we want to ignore the incoming header type. This commit takes care of addressing this issue:

https://github.com/Kong/kong/issues/10246

I am not implmenting the ability to modify the `default_header_type` as opentelemetry notes it should default to `w3c` which it currently is:

https://opentelemetry.io/docs/concepts/signals/traces/#context-propagation

* Update kong/plugins/opentelemetry/schema.lua

---------

Co-authored-by: Samuele Illuminati <samuele@konghq.com>