chore(actions): require manager/lead approval for backport PRs
This commit adds a new workflow that checks if a PR is targeting
`next/**` branches, and the PR is changing any Lua file (except PRs
that changes only tests). If these are found and the PR is not approved
by a manager, then the PR will become red, preventing merges.