fix(queues): fix bug that caused queues to causing queues to not work
Prior to this fix, we always updated ngnix's time stamp before reading it with
nginx.now(), which could result in the timestamp used to determine whether the
coalescing loop still needs to run to be different from the one that was used
to calculate the semaphore timeout.  The latter could become negative due to that
bug, causing entries to be lost.