fix(vault): resource / key parsing problematic with some vaults
### Summary

The vault has generic part, implemented in `kong/pdk/vault.lua` and then
the implementation strategy part implemented in `kong/vaults/*`.

The interface between them is:
```
string, err, ttl = get(conf, resource, version)
```

The generic part supported also a sub-key lookup in case the vault `resource`
returned a multi-value answer as a JSON string. An example of this:

```
❯ PG='{"username":"john","password":"qwerty"}' kong vault get env/pg
{"username":"john","password":"qwerty"}
❯ PG='{"username":"john","password":"qwerty"}' ./bin/kong vault get env/pg/username
john
```

In both examples the `resource` given to Vault `strategy` is `test`.
See how we used the last slash to determine where comes the multi-value lookup key.
For env vault this works fine, but it is problematic with other vaults as we cannot
determine whether `pg/username` is the name of the actual secret or is the `pg` the
name of actual secret, and when it is the first (`pg/username`), we will endup giving
just `pg` to Vault strategy.

This PR changes the character used for multi-value lookups from that `/` to `:`, so example
goes like this:

```
❯ PG='{"username":"john","password":"qwerty"}' kong vault get env/pg
{"username":"john","password":"qwerty"}
❯ PG='{"username":"john","password":"qwerty"}' ./bin/kong vault get env/pg:username
john
```

In AWS Vault the secret name can be:
- `my/secret`, or full ARN like:
- `arn:aws:secretsmanager:us-east-2:123456123456:secret/my/secret-ABCDEFG`

The first form does not allow colons `:`, but the second is filled with colons.

So the key detection function gots from the end of string. If it finds `/`,
it will stop and treat the secret as a single value secret. If it finds `:`
(before finding `/`), it takes the left part from that as a `resource` and
the right part as a `key`. Thus in:

- `my/secret` resource will be `my/secret` and the key is `nil`
- `arn:aws:secretsmanager:us-east-2:123456123456:secret/my/secret-ABCDEFG`
   the resource will be `arn:aws:secretsmanager:us-east-2:123456123456:secret/my/secret-ABCDEFG`
   and the key is `nil`
- `my/secret/another-path:key` resource will be `my/secret/another-path`
   and the key is `key`
- `arn:aws:secretsmanager:us-east-2:123456123456:secret/my/secret-ABCDEFG:key`
   the resource will be `arn:aws:secretsmanager:us-east-2:123456123456:secret/my/secret-ABCDEFG`
   and the key is `key`.

This is a breaking change.

Signed-off-by: Aapo Talvensaari <aapo.talvensaari@gmail.com>