fix(queues): use explicit workspace in registry key (#10721)
This fixes an issue in the queue implementation in EE that made queues fail
to work after the first entry was processed.  This was caused by the 'smart'
way to isolate the queues in each workspace from those in other workspaces by
using __index and __newindex to automatically combine the workspace name into
the key used as index into the table of queues.  This approach failed to work
when the queue had to be deleted as that happens from the timer context in
which no ngx.context is not set.

Fixes KAG-1292