feat(api): node readiness status endpoint (#10610)
This commit introduces a new status API `/status/ready` that can be used to determine
if Kong is in a healthy state and ready to serve user requests. This API returns `200` only
if Kong has a valid configuration and has built the necessary data structures for handling
proxy path requests. It can be used as healthcheck endpoint for load balancers, for example.

KAG-76 KAG-1045

---------

Co-authored-by: Chrono <chrono_cpp@me.com>
Co-authored-by: Harry <harrybagdi@gmail.com>
Co-authored-by: Qi <add_sp@outlook.com>
Co-authored-by: Datong Sun <datong.sun@konghq.com>