feat(cd): upload sbom to nighlies and releases
In light of the new requirement of pushing SBOM to nightlies (in
addition to official releases), this commit replaces the existing
implementation that uses an OSS action to handle tagging, GH release
creation, and asset upload with the `gh(1)` tool. Doing so allows us
more granular control over the steps involved in this process; in
particular, we want to be reusing the same `nightly` tag and the same GH
release for nithgly releases, and the OSS action did not support a "force
push" operation for tags. More generally, one can expect the `gh(1)`
tool gets more eyes and maintenance than the OSS tool, which, with its
simplicity, has over 100 open issues on GH).

FTI-1370