fix(saml): update datafile library to version 0.9-1 (#5268)
The new version of the datafile library fixes a bug that cause Kong not to
start from systemd.  The reason for that was that datafile assumed that the
HOME environment variable would be set.  If it was not set, it would silently
fail to load some of its components and later cause data files required by
the saml plugin not to be found.

Fixes KAG-1466