feat(postgres): add support for using RDS with AWS IAM Authentication (#4518)
This PR adds support for using the RDS database(Postgres engine) with AWS IAM Authentication.

Two new config parameters will be added: pg_iam_auth and pg_ro_iam_auth. When switched on, the Postgres connector will try to generate an IAM authentication token and use the token as the password to connect to the RDS database.

When the IAM authentication method is used, the original pg_password(or pg_ro_password config) will be ignored, and the TLS connection will be enabled mandatorily. The mTLS connection is currently not supported by the RDS database, so the mTLS-related config cannot be configured together with IAM auth.

KAG-167

Co-authored-by: Enrique García Cota <kikito@gmail.com>