fix(queues): display plugin name in queue log entries (#10799)
* fix(queues): display plugin name in queue log entries

Log entries written by queue functions included the name of the queue, but
not the name and ID of the plugin using the queue.  This would make it
difficult to correlate queue log entries written by the queue consumption
side to the configuration.  With this change, an explcit log_tag key can
be passed to the queue when enqueuing which is shown in all log entries
written by the queue.  An explicit log_tag key was chosen over a more
specific 'plugin' or 'plugin_id' key as queues are already used by non-plugin
functionality in EE.

Fixes KAG-1394

(cherry picked from commit b36b5091cd0b337554ebb7a0c9049719769b212b)