chore(ci) stop testing against Cassandra
Cassandra is no longer going to be supported.
This commit removes testing for Cassandra in the CI and will be followed
up with smaller incremental patches to remove all code associated with
Cassandra.
The CI is being disabled first to alleviate CI flakiness and provide a
bit of a relief from CI issues.