refactor: drop requirment to specify cassandra
To support a phased out removal of Cassandra, this commmit removes the
validation for Cassandra. Once all migrations associated with Cassandra
are removed, the schema for migrations will be updated to not allow any
Cassandra fields.
This is done to avoid a large patch that removes all migrations in one
go.