fix(acme): sanity test can work with "kong" storage in Hybrid mode
Currently, there is a technical limitation in hybrid mode, that is, the
DP side cannot perform any write database operation. However, when the
sanity test is executed and the account identifier information is needed
to write to storage, we will skip this operation (after a private
        discussion with @wangchong) and return 200 directly in this
situation. This will not have any substantial impact on the
functionality.