refactor(plugins): drop unused version from plugin rockspec (#5393)
A version bump requires significant work in each release. The most
recent example of this is ef4e9bd4b.

The rockspec file is used by bazel build system to install dependencies
but the version in the rockspec is not used by any other system.
This commit drops versions from plugin-ee rockspec files to minimize
busy work and make it easy to bump up version of Kong after a release.