perf(tests): support zero timeout for logfile assertion
This improves performance for `assert.logfile().has.line()` when the
caller knows that the log file only needs to be read once.

The line-matching logic has also been updated as well. Instead of
reading the entire file at once, we iterate line-by-line and exit early
if the subject is found.