fix(cors): don't send ACAO header where a single origin(not regex) is configured in case of non-matched.
The CORS plugin currently always sets the ACAO header based on the
plugin configuration if the configuration only has a single entry and
contains only non-PCRE metacharacters. This behavior is first introduced
in #2482 and doesn't seem to have any real impact on the functionality I
think.

But this seems is not following the [mozilla
guidelines](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)

> Specifies an origin. Only a single origin can be specified. If the
server supports clients from multiple origins, it must return the origin
for the specific client making the request.

This fixes behavior by no longer sending an ACAO header in this case.