docs(changelog): copy the 3.3.0.0 changelog into master (#5450)
I manually checked the differences.

Almost the changes seemed to be basic mistakes in updating the changelog in master (people would add it to the 3.3 branch instead of to the Unreleased branch, even when committing changes after Code Freeze).

The Exception is :

* Oauth 2.0 Introspection plugin fails when request with JSON that is not a table. [FTI-4974](https://konghq.atlassian.net/browse/FTI-4974)

We *intended* for this one to be in 3.3.0.0, but somehow missed it. I believe we forgot to backport it from master to the 3.3 branch. I have notified victor on the FTI.