refactor(*): drop Cassandra specifics from plugins (#10833) [oss cherry-pick] (#5425)
* refactor(*): drop Cassandra specifics from plugins (#10833)

* refactor: drop requirment to specify cassandra

To support a phased out removal of Cassandra, this commmit removes the
validation for Cassandra. Once all migrations associated with Cassandra
are removed, the schema for migrations will be updated to not allow any
Cassandra fields.
This is done to avoid a large patch that removes all migrations in one
go.

* refactor: drop Cassandra specifics from plugins

This patch is part of a series of patches removing Cassandra support
from Kong. This patch removes all Cassandra specifics from bundled
plugins.

* tests(unit): choose right strategy for `check compat for entities those have `updated_at` field`

---------

Co-authored-by: Qi <add_sp@outlook.com>