tests(wasm): re-number integration test suite
Due to being a long-lived feature branch, several other integration test
directories have been created in master during the lifespan of
feat/wasmx, and there may be others between now and merge time as well.
It's not really important for test suite numbers to be sequential, so
I've bumped us from 15 => 20 in hopes that we won't need to re-number
things again.