tests(helpers): fix failure messages for error/no_error
The primary reason for this commit is to fix some business logic
errors in the eventually() assertion that caused returned/raised
errors/values to be missing from the failure message.

There are a couple additional changes:

It used to be the case that a raised error() would be re-raised if the
caller had not specified ignore_exceptions=true. The intent behind this
was "if the user hasn't told us to expect an error, we should treat it
like any other error and re-raise." However, this means all the nice
context that we've built up is simply dropped, so it's antithetical to
our goal of making things easy to diagnose/debug. Therefore, I've
changed the behavior: unexpected exceptions will now result in a normal
assertion failure message. The raised error itself is included in this
context, so we're not losing any data.

Additionally, I swapped out pcall for xpcall and added a traceback to
the failure message in this case. Because of busted shenanigans, it's
hard to get a particularly meaningful stack trace, but it's better than
nothing.