tests(dbless): use explicit count for worker_processes
To make this test more robust I am adapting it to explicitly configure
the worker process count.

The default value of nginx_main_worker_proceses (spec/kong_tests.conf)
is 1, and there were some assertions in this test suite that relied upon
that default value when checking for the number of common/unique PIDs
reported by the admin API, like so:

assert.equal(1, nonmatching)

Updating that value from 1 to WORKER_PROCS is not sufficient, because
the business logic for calculating the number of common/unique PIDs
effectively double-counted unique PIDs, so WORKER_PROCS => 4 would mean
a value of 16 for nonmatching.

So in order to make the test work with any arbitrary number for
WORKER_PROCS, I had to update the PID-counting logic. As a result the
tests are much more explicit and will result in more meaningful error
messages if they fail.